chore: 회원에 대한 파일들의 패키지 이동
