+++ src/main/java/woowacourse/auth/ui/AuthController.java
@@ -2,16 +2,13 @@ package woowacourse.auth.ui;
 
 import javax.validation.Valid;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import woowacourse.auth.application.AuthService;
 import woowacourse.auth.dto.request.LoginRequest;
 import woowacourse.auth.dto.request.PasswordCheckRequest;
-import woowacourse.auth.dto.response.EmailUniqueCheckResponse;
 import woowacourse.auth.dto.response.LoginResponse;
 import woowacourse.auth.dto.response.PasswordCheckResponse;
 import woowacourse.auth.support.AuthenticationPrincipal;
