+++ src/main/java/woowacourse/auth/ui/AuthControllerAdvice.java
@@ -7,20 +7,16 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import woowacourse.auth.dto.response.ErrorResponse;
+import woowacourse.shoppingcart.dto.response.ErrorResponse;
 import woowacourse.auth.exception.AuthorizationException;
 
 @RestControllerAdvice
 public class AuthControllerAdvice {
 
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<ErrorResponse> handleInvalidRequest(IllegalArgumentException e) {
-        return ResponseEntity.badRequest().body(new ErrorResponse(e.getMessage()));
-    }
-
     @ExceptionHandler(AuthorizationException.class)
     public ResponseEntity<ErrorResponse> handleUnauthorizedRequest(AuthorizationException e) {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ErrorResponse(e.getMessage()));
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body(new ErrorResponse(e.getMessage()));
     }
 
     @ExceptionHandler(BindException.class)
@@ -30,6 +26,7 @@ public class AuthControllerAdvice {
                 .stream()
                 .map(DefaultMessageSourceResolvable::getDefaultMessage)
                 .collect(Collectors.joining(System.lineSeparator()));
-        return ResponseEntity.badRequest().body(new ErrorResponse(exceptionMessage));
+        return ResponseEntity.badRequest()
+                .body(new ErrorResponse(exceptionMessage));
     }
 }
