+++ src/main/java/woowacourse/shoppingcart/dao/MemberDao.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.dao;
+package woowacourse.shoppingcart.dao;
 
 import java.util.Optional;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import woowacourse.auth.domain.Member;
+import woowacourse.shoppingcart.domain.value.Member;
 
 @Repository
 public class MemberDao {
