+++ src/main/java/woowacourse/shoppingcart/domain/value/Nickname.java
@@ -1,23 +1,22 @@
-package woowacourse.auth.domain;
+package woowacourse.shoppingcart.domain.value;
 
 import java.util.Objects;
 import java.util.regex.Pattern;
 
-public class Email {
+public class Nickname {
 
-    private static final Pattern PATTERN = Pattern.compile("^(?=.{1,64}@)[A-Za-z0-9_-]"
-            + "+(\\.[A-Za-z0-9_-]+)*@[^-][A-Za-z0-9-]+(\\.[A-Za-z0-9-]+)*(\\.[A-Za-z]{2,})$");
+    private static final Pattern PATTERN = Pattern.compile("^[가-힣]{1,5}$");
 
     private final String value;
 
-    public Email(String value) {
+    public Nickname(String value) {
         validate(value);
         this.value = value;
     }
 
     private void validate(String value) {
         if (!PATTERN.matcher(value).matches()) {
-            throw new IllegalArgumentException("이메일 형식이 올바르지 않습니다.");
+            throw new IllegalArgumentException("닉네임 형식이 올바르지 않습니다.");
         }
     }
 
@@ -33,8 +32,8 @@ public class Email {
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        Email email = (Email) o;
-        return Objects.equals(value, email.value);
+        Nickname nickname = (Nickname) o;
+        return Objects.equals(value, nickname.value);
     }
 
     @Override
