+++ src/main/java/woowacourse/shoppingcart/domain/value/Password.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.domain;
+package woowacourse.shoppingcart.domain.value;
 
 import java.util.Objects;
 import java.util.regex.Pattern;
