+++ src/main/java/woowacourse/shoppingcart/dto/request/MemberCreateRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.dto.request;
+package woowacourse.shoppingcart.dto.request;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.Pattern;
