+++ src/main/java/woowacourse/shoppingcart/dto/request/PasswordUpdateRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.dto.request;
+package woowacourse.shoppingcart.dto.request;
 
 import javax.validation.constraints.Pattern;
