+++ src/main/java/woowacourse/shoppingcart/dto/response/EmailUniqueCheckResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.dto.response;
+package woowacourse.shoppingcart.dto.response;
 
 public class EmailUniqueCheckResponse {
