+++ src/main/java/woowacourse/shoppingcart/ui/ControllerAdvice.java
@@ -9,7 +9,7 @@ import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import woowacourse.auth.dto.response.ErrorResponse;
+import woowacourse.shoppingcart.dto.response.ErrorResponse;
 import woowacourse.shoppingcart.exception.InvalidCartItemException;
 import woowacourse.shoppingcart.exception.InvalidCustomerException;
 import woowacourse.shoppingcart.exception.InvalidOrderException;
@@ -23,6 +23,12 @@ public class ControllerAdvice {
     public ResponseEntity<ErrorResponse> handleUnhandledException() {
         return ResponseEntity.internalServerError()
                 .body(new ErrorResponse("서버에 문제가 발생했습니다."));
+    }
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity<ErrorResponse> handleInvalidRequest(IllegalArgumentException e) {
+        return ResponseEntity.badRequest()
+                .body(new ErrorResponse(e.getMessage()));
     }
 
     @ExceptionHandler({MethodArgumentNotValidException.class})
