+++ src/test/java/woowacourse/shoppingcart/application/CartServiceTest.java
@@ -9,8 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-import woowacourse.auth.application.MemberService;
-import woowacourse.auth.dto.request.MemberCreateRequest;
+import woowacourse.shoppingcart.dto.request.MemberCreateRequest;
 import woowacourse.shoppingcart.dto.request.CartItemRequest;
 import woowacourse.shoppingcart.dto.request.ProductRequest;
 import woowacourse.shoppingcart.dto.response.CartItemResponse;
