+++ src/test/java/woowacourse/shoppingcart/domain/value/EmailTest.java
@@ -1,10 +1,11 @@
-package woowacourse.auth.domain;
+package woowacourse.shoppingcart.domain.value;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import woowacourse.shoppingcart.domain.value.Email;
 
 class EmailTest {
